Version Packages (#26)
This PR was opened by the [Changesets release](https://github.com/changesets/action) GitHub action. When you're ready to do a release, you can merge this and publish to npm yourself or [setup this action to publish automatically](https://github.com/changesets/action#with-publishing). If you're not ready to do a release yet, that's fine, whenever you add more changesets to main, this PR will be updated.


# Releases
## ctw-component-library@0.0.2

### Patch Changes

-   54e83e4: Adds the ability to add changeset to this repo and adds MIT license

Co-authored-by: github-actions[bot] <github-actions[bot]@users.noreply.github.com>