Version Packages (#30)
This PR was opened by the [Changesets release](https://github.com/changesets/action) GitHub action. When you're ready to do a release, you can merge this and publish to npm yourself or [setup this action to publish automatically](https://github.com/changesets/action#with-publishing). If you're not ready to do a release yet, that's fine, whenever you add more changesets to main, this PR will be updated.


# Releases
## @zus-health/ctw-component-library@0.0.3

### Patch Changes

-   1fad275: fix build script
-   1fad275: Update readme and adding changeset to set process

Co-authored-by: github-actions[bot] <github-actions[bot]@users.noreply.github.com>