Work around for theme typing
The difficulty here is due to the tailwind config using commonjs and not ES Modules. We need to import our theme in both the tailwind config commonjs file AND in our TypeScript/ES Module provider file. The work around is to use TS as the base and compile it to a cjs file for tailwind config to use.