Ctw 283/edit add conditions (#56)
Creating a draft pull request, not yet ready for review but will mark it
as such when ready.

Checklist:

- [x] Add condition from lens
- [x] Add empty condition
- [x] Use env vars for system and patientID instead
- [x] Fix dropdown styles
- [x] Prefill conditions with as much data as available
- [x] Possible refactors to cleanup how code is structured (lots of
moving parts, might be able to simplify it)
- [x] Add Changeset
- [x] Edit condition
- [x] Add case for hidden fields
- [x] Fix typescript issues

**Commments**
- AC doesnt show SubjectID as a field, but being that we are using its
value to post the condition, should we show it but add it as locked?
- PatientProvider conondrum. Currently it always returns a patientModel,
but we probably need to change to handle the case where we pass a UPID.
Do we then just return a string and have it return string or model? Or
do we just always make that request? Or do we figure out to see if we
can cache/store the patientModel?

**In Future PR**
- [ ] Transform hex colors to RGBA (different story?)
- [ ] Update conditions list when posting/patching