move context out of provider file to avoid refresh bug (#95)
Moving the CTW Context and provider into separate files to avoid issues
when changing certain files it errors out and says context is not
defined. This is caused by a bug in vite which there is a current open
pr on.. https://github.com/vitejs/vite/pull/10239, which should solve
this. Being that this bug is annoying enough in dev i am splitting
context and provider into separate files to avoid this bug.