Ctw 392/fix patient overfetching (#101)
This pr: 

- Ensures that it only fetches when on the conditions component and not
elsewhere.
- We used to fetch everywhere because our patient provider is used in
the top level of apps and when patientID, systemUrl or PatientUPID
changed it fetched again, irregardless of whether they were in our
conditions component. This is behavior, in my opinion we want to avoid.
We should try to keep fetching behavior tied down to our components to
avoid weird behavior like this. By switching to React Query and calling
it in the component we stop that behavior, are able to cache calls to
get patients, and if we reuse the same query anywhere else... say
medications it will use the cache instead of making a new call.

Reasoning behind certain changes: 
- Only returning system URL and patientID from the provider. This way
anywhere we usePatient we dont have to handle the case of UPID as our
patient provider now handles that in the patientProvider. Only confusion
could be that that it is called patientID instead of patientUPID, as it
could technically be both. Im open for debate on this, and or
suggestions.