Refactor useQuery usage and provide more context
We now pass around a CTWRequestContext which can hold a number of things useful for fetching data. This now includes the builderId (via claims data or CTWProvider property).

Refactored uses of useQuery to hopefully simplify things a bit via a new `useQueryWithPatient` which will take care of getting the requestContext and patient model for us.