CTW-432/zus account builder (#142)
* Add optional `builderId` prop to `CTWProvider`. When set, we'll use
that builderId for any filtering AND we'll set the `Zus-Account` header
for FHIR requests.
* Refactor `CTWProvider` and search code to pass around a new
`CTWRequestContext`. This gives us one place to add things that any of
our requests need. For now that consists of `env, builderId, authToken,
fhirClient`.
* Add new `useQueryWithPatient` helper which will handle setting up the
`patient` model and the `requestContext`. Refactor existing uses of
`useQuery` to use that instead.
* Add a default `staleTime` of `30 seconds` to all `useQuery` queries.
This means that we will use the cached result for queries based on their
`keys`.
* Add a `staleTime` of `5 minutes` for patient queries.