CTW-491/fix condition autocomplete (#179)
This fixes a bug where our condition autocomplete was grabbing the
`authToken` from `getCTW`. This auth token was only set if the app
passed in `authToken` to `CTWProvider`. Instead, everything should go
through the `getRequestContext` which will correctly grab and ensure a
non-expired authToken.

Additionally:
* Persist autocomplete search event as the use of debounce causes react
to garbage collect the event.
* Add an `isLoading` to fix a bug where autocomplete would briefly say
no results found while the search was running.