Export drawer from index (#203)
Exporting parts of the library that would be used if someone were
building custom components using the library. Currently it is difficult
to build a custom medication table and drawer like `<PatientMedications
/>` because we don't export some of the useful building blocks of the
library.

It's possible that future refactors might make some of these exports
obsolete, but if that's the case we should just version bump
accordingly.