CTW-349 Capitalizing fields in details/history drawer (#211)
Capitalizing various fields in the details/history drawers. I figured
there was 3 options here:

1. Do it generically in the underlying core components with something
like a `capitalize` boolean prop
2. Do it in the getters
3. Do it when setting up data to pass to the component for presentation

I chose option 3 because it felt like this kind of decision belongs
close to the presentation layer. The underlying data may not be
capitalized, and this is likely a case-by-case decision (eg. we're not
going to capitalize the "h" at the beginning of a URL), so I thought
this was the appropriate place to put this logic.