CTW-526/Create ActionList (match canvas UI) (#216)
Creates an `<ActionList/>` component which will look something like
this:
<img width="1205" alt="Screen Shot 2022-11-09 at 12 18 44 PM"
src="https://user-images.githubusercontent.com/6380075/200897404-c95a8a99-308d-461d-9c96-19ff4dd6bf24.png">

And in the wild:
<img width="544" alt="Screen Shot 2022-11-09 at 12 27 48 PM"
src="https://user-images.githubusercontent.com/6380075/200899296-f7394e81-7781-4c1b-b591-28db4c58c1ad.png">

Notable attributes:
List items marked "active" will show a (primary) colored border to the
left by default and when hovered present a button to take action on that
item. Optionally the opposite can be done for inactive items if an
`onUndoAction` handler is passed in, but that is not a requirement.

It is the duty of the data owner to set the "active" property on their
item through their custom "onAction" handler. The list is simply
rendering state and wiring up the event listeners.

There is an optional "onRowClick" handler that will take action if the
row is clicked or enter button is pushed while row is focused. It is the
responsibility of the code-base utilizing this component to create a
style for the focus state using CSS. We are not implementing focused css
styles or class-names at the time. The keyboard handler here is less
about accessibility than it is about ensuring we're not actively
preventing consumers of the library from building accessible
applications using Zus components.