CTW-546/Prescriber sometimes blank (#256)
https://zeushealth.atlassian.net/browse/CTW-546

When MedicationDrawer shows details of a medication it usually has a
blank prescriber. This is because the details come from a
MedicationStatement which only knows about a prescriber if the
`"https://zusapi.com/lens/extension/medicationLastPrescriber"` lens is
set.

It is usually (if not always) the case that this lens is not set and so
the field is blank in the details ui. Now the component will check if
that field is blank again after fetching related medication* resources
from ODS. It will then find the latest MedicationRequest (and included
requester resource) and use those to fill in the prescriber field.