ActionList doesn't always need action handler (#291)
ActionList had `onAction` as required (at time of creation it made
sense) but we found a use-case where the list was used without the
action and had to be written as:
```tsx
              <ActionList
                className="ctw-medication-review-list"
                items={actionItems(builderMedications)}
                onAction={() => {}}
                onRowClick={({ medication }) => openMedicationDrawer(medication)}
                actionText=""
              />
```
so this makes `actionText` and `onAction` default to those empty values.

Additionally added `.ctw-drawer-title` to the drawers h2 tag so it can
be easily targeted by CSS rules. We found a case where semanticui was
messing up the h2 style in an app using ctw-component-library. We don't
want to restyle from within, so this classname just makes it easier to
target from without.