CTW-569/archive other conditions (#306)
Add `Archive` and `Un-Archive` menu options to other provider records
"..." menu. Archiving for the first time will create a new `Basic`
resource and associate the resource with the summary condition and the
user performing the action. Subsequent un-archiving/re-archiving is done
by updating the existing basic resource with the proper meta tags AND
the current user's info.

Additional changes:
* Add ESLint rule to prevent unused imports.
* Refactored how we fetch practitioner references.
* Delete modal now shows disabled remove button with spinner.