Adding UPID to CPR query (#347)
When retrieving the history for a given medication we look up a list of
medicationX resources in ODS by ID.

In order to trigger "CPR mode" in ODS, where we surface data across
builders, we must search by UPID. This was thought to be a
redundant/unnecessary filter, but it is actually required in order to
access the data. It's also worth noting that the CPR data is also only
available via a search (ie. using the `_id` query parameter) and is not
accessible directly by ID.

This does not impact conditions as the condition history fetch is
already passing a UPID (see line 122 of `conditions.ts`).