CTW-686 Updating logic for getting the identifying RxNorm code on a medication (#410)
Our logic to get the identifying RxNorm code on a medication statement
should not have been preferring generics (ie. SCD RXCUIs). This has been
causing an issue with duplicating resources via the reconciliation
workflow in Canvas - when a user is presented with a branded drug (ie.
an underlying SBD RXCUI) we are using the SCD to look up the medication
in Canvas and thus creating a new unique lens resource after the
reconciliation process.