CTW-765 pushable issue (#495)
In some EHRs we have this issue where opening a CTW drawer causes the
entire chart to shift. We had previously gotten around this by adding
logic to temporarily remove this `pushable` class from the chart just
while the drawer is opening. Now that we've reduced our footprint by
simply dropping in our `PatientMedicationsTabbed` component we no longer
have the flexibility to add this kind of workaround when opening our
drawer.

This PR exposes some props on the medications component that can be used
to execute a function when opening the history drawer.

Longer term we should find a way to more generally expose these kinds of
functions to run upon opening a drawer because they are obviously not
specific to medications.


https://user-images.githubusercontent.com/97455910/218780622-83e3853c-ab07-4544-b63f-9fba17beefac.mov

---------

Co-authored-by: Michael Rosata <mrosata@zushealth.com>