Version Packages (#653)
This PR was opened by the [Changesets
release](https://github.com/changesets/action) GitHub action. When
you're ready to do a release, you can merge this and the packages will
be published to npm automatically. If you're not ready to do a release
yet, that's fine, whenever you add more changesets to main, this PR will
be updated.


# Releases
## @zus-health/ctw-component-library@1.3.2

### Patch Changes

-   f2e2e977: Fix issue where allergies were not sorted correctly
- d0610b37: Export patient history table. Add handleClickRow handler to
table.
-   f62ec0e4: Fix last retrieved date spacing.

Co-authored-by: github-actions[bot] <github-actions[bot]@users.noreply.github.com>