Version Packages (#667)
This PR was opened by the [Changesets
release](https://github.com/changesets/action) GitHub action. When
you're ready to do a release, you can merge this and the packages will
be published to npm automatically. If you're not ready to do a release
yet, that's fine, whenever you add more changesets to main, this PR will
be updated.


# Releases
## @zus-health/ctw-component-library@1.3.7

### Patch Changes

-   5b1c0eb9: Add source document link to observation drawer
- 93406b4c: Update search helpers to support Zus refactoring to store
3rd party data in dedicated builders.
-   17e2143a: Make condition display value bold.

Co-authored-by: github-actions[bot] <github-actions[bot]@users.noreply.github.com>