CTW-1006 Updated metrics (#678)
This PR adds a few metrics:

- `component.{component name}.failure` will indicate instances of a
component failing to load. This will help us report out on how many
times a user got an error when trying to pull up the ZAP.
- `component.{component name}.loaded` will indicate how many times a
user pulled up the ZAP. This may be a noisy metric as this is based on
`componentDidMount`, so it's possible the component was loaded but
wasn't actually visible to the user. But this is a good start.
- `records.{record type}` (eg. `records.outside_medications`) will
capture how many records were loaded for a given component in the ZAP.
We'll use this to capture things like "on average users see N outside
medications in the ZAP" and will also help to identify the number of
instances where the ZAP was empty for a given resource type. Right now
this PR only adds that metric for medications. If there's no opposition
to this approach I'll add the rest.

These metrics are sent to a Zus endpoint and not directly to DataDog
from the client. Our plan moving forward is to remove the RUM telemetry
from this repo (which requires opting in to enable telemetry) and just
use that for the standalone Zus app.

---------

Co-authored-by: Michael Rosata <mrosata@zushealth.com>