CTW-1069 Fixing queries that need to filter by users builder ID (#704)
We removed the logic that filters by the users builder ID because we
believed that the firstparty search parameter would take care of that
for us. But that search parameter doesn't filter out all other builders
data, it just filters out third party data. So because of CPR our query
still returns data from other builders.

---------

Co-authored-by: Michael Rosata <mrosata@zushealth.com>