Version Packages (#714)
This PR was opened by the [Changesets
release](https://github.com/changesets/action) GitHub action. When
you're ready to do a release, you can merge this and the packages will
be published to npm automatically. If you're not ready to do a release
yet, that's fine, whenever you add more changesets to main, this PR will
be updated.


# Releases
## @zus-health/ctw-component-library@1.7.10

### Patch Changes

- 4094dbd6: Add version to metrics so we can track what versions are
being used.

Co-authored-by: github-actions[bot] <github-actions[bot]@users.noreply.github.com>