Fix sorting & display of patient history list (#717)
CTW-1088

The patient history API returns data sorted by targetDate descending and
has no sorting options. We display the createdAt field. This causes
confusion as to why the table is ordered the way it is - for a customer
that uses appointment-based patient history automation we are setting a
target date based on upcoming appointments, so we're showing a bunch of
unresolved/upcoming patient history requests on the first page, and they
have very disparate create timestamps, so the data looks kind of
nonsensical. We can alleviate this for the time being by displaying the
target date for the "Last Queried" date and filtering the search to only
return patient history requests that have a target date up until the
current date (ie. don't show those that are set to happen in the
future).

<img width="984" alt="image"
src="https://github.com/zeus-health/ctw-component-library/assets/16874444/7e8acaaa-c455-416a-b943-d5f59d5627d2">

---------

Co-authored-by: Luthfi Bustillos-Francis <luthfibustillos@gmail.com>
Co-authored-by: Michael Rosata <mrosata@zushealth.com>