Version Packages (#734)
This PR was opened by the [Changesets
release](https://github.com/changesets/action) GitHub action. When
you're ready to do a release, you can merge this and the packages will
be published to npm automatically. If you're not ready to do a release
yet, that's fine, whenever you add more changesets to main, this PR will
be updated.


# Releases
## @zus-health/ctw-component-library@1.8.0

### Minor Changes

-   fcde05f4: Add a beta label to Care Team tab in ZAP

Co-authored-by: github-actions[bot] <github-actions[bot]@users.noreply.github.com>