[BUG] Loading State for Dismiss/Restore Buttons (#695)
When the dismiss/restore button is clicked for patient outside
conditions and medications, the button shows a loading state and the add
button is disabled until the action is complete.

- Create `useAction` hook to help with chaining callbacks
- Create `useToggleArchive` hook to help with providing input feedback
to the user when the action is invoked
- Add styling for disabled primary and secondary buttons
- Update the patient outside conditions and medications components to
provide feedback to the user when the dismiss/restore buttons are
clicked. When it is clicked, both the dismiss/restore and add buttons
are disabled until the action is completed. If it is successful, then
the query is invalidated and rerun

<img width="597" alt="image"
src="https://user-images.githubusercontent.com/11450375/235786160-3070b063-1291-40a6-b192-e6aebeb082e0.png">

---------

Co-authored-by: Priyank Madria <98342697+pmadria3@users.noreply.github.com>