Version Packages (#775)
This PR was opened by the [Changesets
release](https://github.com/changesets/action) GitHub action. When
you're ready to do a release, you can merge this and the packages will
be published to npm automatically. If you're not ready to do a release
yet, that's fine, whenever you add more changesets to main, this PR will
be updated.


# Releases
## @zus-health/ctw-component-library@1.12.0

### Minor Changes

- 57e03269: Added `fhirWriteBackClient` to proxy creation of new FHIR
resources through the ehr-data-integration server.

### Patch Changes

-   88c8d13c: Update some dependencies
-   bdc77c6f: Fix case where CCDA viewer crashes

Co-authored-by: github-actions[bot] <github-actions[bot]@users.noreply.github.com>