CTW-1185/fqs read your writes (#791)
Setup long polling from FQS in our generic `createOrEditFhirResource`
helper. We'll need to be very careful when introducing other helpers
that mutate data in ODS. Hopefully we can keep using this existing
helper which now await's FQS to have the updated data.

Additionally fixed a bug where we were not creating a provenance
resource for created conditions and basic resources.