CTW-1186/onResourceSave optional callback (#793)
This feature allows an optional callback function `onResourceSave` that
providers can pass in from on the provider context. Whenever a resource
gets created or edited in the function `createOrEditFHIRResource` the
callback is fired. The backstory is that some providers wish to be
notified in the frontend after a save or edit is performed so they can
perform work as well.
---------

Co-authored-by: Wesley Reitzfeld <1568246+WesleyKapow@users.noreply.github.com>