Version Packages (#861)
This PR was opened by the [Changesets
release](https://github.com/changesets/action) GitHub action. When
you're ready to do a release, you can merge this and the packages will
be published to npm automatically. If you're not ready to do a release
yet, that's fine, whenever you add more changesets to main, this PR will
be updated.


# Releases
## @zus-health/ctw-component-library@1.23.1

### Patch Changes

- 5706ff6d: Refactored FQS vs ODS query logic and removed all "drawer"
specific feature variants. We'll instead use resource specific variants,
regardless of which components do the data fetching.
- 7bb72d33: Fetch timeline data (encounters, diagnostic reports,
observations) from FQS. This is behind a feature flag, default is to
still fetch from ODS.

Co-authored-by: github-actions[bot] <github-actions[bot]@users.noreply.github.com>