Version Packages (#864)
This PR was opened by the [Changesets
release](https://github.com/changesets/action) GitHub action. When
you're ready to do a release, you can merge this and the packages will
be published to npm automatically. If you're not ready to do a release
yet, that's fine, whenever you add more changesets to main, this PR will
be updated.


# Releases
## @zus-health/ctw-component-library@1.23.2

### Patch Changes

- 7297d2fc: Add `unleash-proxy-client` to deps. It was previously listed
as peer dep.

Co-authored-by: github-actions[bot] <github-actions[bot]@users.noreply.github.com>