Version Packages (#866)
This PR was opened by the [Changesets
release](https://github.com/changesets/action) GitHub action. When
you're ready to do a release, you can merge this and the packages will
be published to npm automatically. If you're not ready to do a release
yet, that's fine, whenever you add more changesets to main, this PR will
be updated.


# Releases
## @zus-health/ctw-component-library@1.23.3

### Patch Changes

- 860bb086: Alter filtering for allergies client side instead of server
side temporarily.

Co-authored-by: github-actions[bot] <github-actions[bot]@users.noreply.github.com>