Version Packages (#874)
This PR was opened by the [Changesets
release](https://github.com/changesets/action) GitHub action. When
you're ready to do a release, you can merge this and the packages will
be published to npm automatically. If you're not ready to do a release
yet, that's fine, whenever you add more changesets to main, this PR will
be updated.


# Releases
## @zus-health/ctw-component-library@1.24.0

### Minor Changes

- 4d77b89b: Add managing organization to medication statement FQS
queries
- 45c813ee: Use summary tag in medication lens query instead of
ActiveMedication tag

### Patch Changes

- a0e46a38: Enable fetching provenances from FQS if "provenances"
feature-flag variant is enabled.
-   e10b3e9a: Don't filter out resources without dates

Co-authored-by: github-actions[bot] <github-actions[bot]@users.noreply.github.com>