CTW-1240 Distinguish builder and outside data in allergy component (#852)
This PR primarily focuses on the following improvements to the allergy
component:

- Better de-duping of builder records vs outside records. Builder
records are prioritized over outside records and will be displayed with
a check icon.
- Outside records now have a "Dismiss" button that works the same as it
does for medications and conditions.
- Outside records also have a "Mark as Viewed" button that makes an
outside record go from bold to normal font. Clicking on an outside
record to view the details also has this effect. This is the first step
towards having this functionality in other components.

There's also some general refactoring around re-naming "archive" to
"dismiss" in a bunch of places to better match the current product
terminology.


https://github.com/zeus-health/ctw-component-library/assets/97455910/6734fa58-38f4-479e-b826-9462d19bce26

---------

Co-authored-by: Wesley Reitzfeld <1568246+WesleyKapow@users.noreply.github.com>