Version Packages (#888)
This PR was opened by the [Changesets
release](https://github.com/changesets/action) GitHub action. When
you're ready to do a release, you can merge this and the packages will
be published to npm automatically. If you're not ready to do a release
yet, that's fine, whenever you add more changesets to main, this PR will
be updated.


# Releases
## @zus-health/ctw-component-library@1.26.2

### Patch Changes

- 3bd05c15: Distinguish builder and outside data in allergy component.
Add "Dismiss" and "Mark as Read" functionality to allergy component.

Co-authored-by: github-actions[bot] <github-actions[bot]@users.noreply.github.com>