fixing styling of notification icons (#901)
Right now the notification icon doesn't always display to the left, and
vertically aligned, with the tab title.

We do still have an issue with the tab title content shifting when the
notification icon goes away (ie. you've read the last resource). We
could address this by having the icon always there and adjusting the
opacity, but that looks pretty awkward with the icon on the left.
Perhaps if/when we move the icon to the right we can take this approach.

<img width="500" alt="image"
src="https://github.com/zeus-health/ctw-component-library/assets/97455910/4ea152c6-ace9-4bf2-93b2-46b255f6a315">