POC for writing documents directly to EHR API (#832)
Exposes an `onAddToRecord` prop for `PatientDocuments` that accepts a
callback function that takes a `DocumentModel` and `fhir4.Binary`. If
that prop is set then we show an "Add" button on the documents component
that executes the callback function. Before executing the callback it
retrieves the binary for the document. Ultimately this can be used by
consumers to trigger a post to an endpoint in their API that imports
documents.

I think an interesting longer term option here may be a more generic
prop called `additionalRowActions` that takes an array of callback
functions and names (for the buttons) so that consumers can trigger
whatever they want for an arbitrary number of actions. But for now I
plan to leave this as a single onAddToRecord prop.

---------

Co-authored-by: Michael Rosata <mrosata@zushealth.com>