Version Packages (#923)
This PR was opened by the [Changesets
release](https://github.com/changesets/action) GitHub action. When
you're ready to do a release, you can merge this and the packages will
be published to npm automatically. If you're not ready to do a release
yet, that's fine, whenever you add more changesets to main, this PR will
be updated.


# Releases
## @zus-health/ctw-component-library@1.27.9

### Patch Changes

- 63ce552f: Forcing medication history results to be sorted by resource
ID before de-duplicating.
-   06196609: Add not authorized error message.

---------

Co-authored-by: github-actions[bot] <github-actions[bot]@users.noreply.github.com>
Co-authored-by: Patrick Valle <github@patrickvalle.com>