CTW-1283 Moving view FHIR to quick actions for allergies & immz (#921)
Modified allergies & immunizations so that the View FHIR button is in
the row actions and not an always present column. Will make similar
modifications to other tables as we bring them into alignment with the
read/unread functionality.

I had considered doing this as generic functionality in `ResourceTable`
but was not confident in how to do it without a somewhat major refactor.
The `rowActions` that gets passed down is a `JSX.Element` that contains
the contents of the row actions that get overlayed on the row content. I
can't see a clean way of injecting the "View FHIR" button into that
element. One option could've been to modify rowActions to take in an
array of actions, and construct the row actions from that array, but
then that takes away some of the flexibility that rowActions affords in
its current state. I'm also not confident that it's a given that every
`ResourceTable` should necessarily show the View FHIR button.



https://github.com/zeus-health/ctw-component-library/assets/97455910/1c21feed-ec98-4b2b-9fb4-126b0f9ea32a