CTW-1279 New PatientMedicationsAll component (#954)
Added new `PatientMedicationsAll` component which will replace the
existing medications components in standalone. The existing components
should be considered deprecated now but we will need to work with
individual consumers to migrate off them.

Note that medications are lacking all CRUD operations like conditions
have - you cannot add a new medication and you cannot edit/remove
builder-owned medications. All you can do is add a medication from the
outside list to your builder. Getting to parity with conditions is not
in scope for this effort and that is approved by product/UX.
---------
Co-authored-by: Michael Rosata <mrosata@zushealth.com>