Version Packages (#966)
This PR was opened by the [Changesets
release](https://github.com/changesets/action) GitHub action. When
you're ready to do a release, you can merge this and the packages will
be published to npm automatically. If you're not ready to do a release
yet, that's fine, whenever you add more changesets to main, this PR will
be updated.


# Releases
## @zus-health/ctw-component-library@1.32.0

### Minor Changes

-   b40d007c: Expand empty table display to include requesting records
-   783ce18c: Update storybook tests to include FQS queries.

### Patch Changes

- fd0c38fe: Update condition history details and history card null
states.
-   ce476a80: Change default colors to use Zus theme.

Co-authored-by: github-actions[bot] <github-actions[bot]@users.noreply.github.com>