Version Packages (#975)
This PR was opened by the [Changesets
release](https://github.com/changesets/action) GitHub action. When
you're ready to do a release, you can merge this and the packages will
be published to npm automatically. If you're not ready to do a release
yet, that's fine, whenever you add more changesets to main, this PR will
be updated.


# Releases
## @zus-health/ctw-component-library@1.34.0

### Minor Changes

-   2ac573b7: Added trend data to Glucose and A1C labs

### Patch Changes

- e9b13066: Remove role field from dropdown and default to SNOMED
158965000 for role.

Co-authored-by: github-actions[bot] <github-actions[bot]@users.noreply.github.com>