Version Packages (#1001)
This PR was opened by the [Changesets
release](https://github.com/changesets/action) GitHub action. When
you're ready to do a release, you can merge this and the packages will
be published to npm automatically. If you're not ready to do a release
yet, that's fine, whenever you add more changesets to main, this PR will
be updated.


# Releases
## @zus-health/ctw-component-library@1.37.0

### Minor Changes

- bba15492: Updated FQS Observations query to include filtering by code

### Patch Changes

-   34f1deee: Set default resources for ZAP.
-   56aaa434: Make unknown condition type display as "Unknown".
- 0a8aa100: Improve responsiveness of ZAP tabs to collapse overflowed
tabs into a more menu.

Co-authored-by: github-actions[bot] <github-actions[bot]@users.noreply.github.com>