CDEV-296: Only fetch trend data if Patient is new (#1007)
https://zeushealth.atlassian.net/browse/CDEV-296

The FQS "filter Observations by LOINC code" feature is now available,
but unfortunately only net-new Observations will be returned until a
backfill of all historical data is complete. This means that we could
possibly return incomplete trend data if a Patient has historical
Observations **and** net-new Observations.

To guard against returning incomplete trend data, this PR checks when
the Patient was created and will only fetch Observations when the
Patient was created after the 7/19/2023 date (which is when FQS first
introduced the filter feature).

This means we will only show trend data on recently created Patients.
Once the FQS backfill of Observation data is complete, we can remove
this restriction and show trend data for all Patients.