Version Packages (#1018)
This PR was opened by the [Changesets
release](https://github.com/changesets/action) GitHub action. When
you're ready to do a release, you can merge this and the packages will
be published to npm automatically. If you're not ready to do a release
yet, that's fine, whenever you add more changesets to main, this PR will
be updated.


# Releases
## @zus-health/ctw-component-library@1.41.0

### Minor Changes

-   271a3355: Added user_id to telemetry metrics

### Patch Changes

-   2fd3affc: Add click count for diagnostic reports.
-   93d86d3c: Cleanup of telemetry namespace dataset attributes

Co-authored-by: github-actions[bot] <github-actions[bot]@users.noreply.github.com>