Don't wait for mark as viewed when opening drawers
Reworked dismiss/restore and read/new to update the model optimistically. This fixes a bug where the row action button on hover would briefly show the old state after the spinner was done. It also allows us to mark a record as read and not wait before opening the drawer. The drawer will then correctly render a "mark as new" button and not a "mark as viewed" button.