CDEV-70/details drawer actions (#1022)
Rework row actions and plumb them to the details drawer. This way
details drawer will always show the same actions as shown on row hover.

Additionally:
* Rework row actions to optionally take an `onSuccess` prop so that we
can close the drawer for certain actions (dismiss/mark as new/remove).
* Refactor the addition of dismiss & read actions into
`useAdditionalResourceActions` so that it can be used by both the
`ResourceTable` and `ResourceDetailsDrawer`.
* Rework dismiss/restore and read/new to update the model record
optimistically. This fixes a bug where the row action button on hover
would briefly show the old state after the spinner was done. It also
allows us to mark a record as read and not wait before opening the
drawer. The drawer will then correctly render a "mark as new" button and
not a "mark as viewed" button.
* Rework rendering of row actions to return null if there are no row
actions. This allows the details drawer footer to be hidden when there
are zero row actions.
* Details drawer was previously taking `enableFQS` but not doing
anything with it. Removed!