Fix capitalization for resource table title/description column (#1044)
Our default behavior was to capitalize the resource table title but
that's really a very case-by-case decision. Labs/immz/meds all have
strings that don't make sense when you generically apply the lodash
capitalize function. So this PR removes that behavior from the title
component and leaves it up to the model getter to decide how to
capitalize.