Version Packages (#1045)
This PR was opened by the [Changesets
release](https://github.com/changesets/action) GitHub action. When
you're ready to do a release, you can merge this and the packages will
be published to npm automatically. If you're not ready to do a release
yet, that's fine, whenever you add more changesets to main, this PR will
be updated.


# Releases
## @zus-health/ctw-component-library@1.45.4

### Patch Changes

-   2a5b770e: Fix capitalization logic for resource table titles.
-   0875afdd: Fix show all for the diagnostic report.

---------

Co-authored-by: github-actions[bot] <github-actions[bot]@users.noreply.github.com>
Co-authored-by: Patrick Valle <github@patrickvalle.com>