Version Packages (#1046)
This PR was opened by the [Changesets
release](https://github.com/changesets/action) GitHub action. When
you're ready to do a release, you can merge this and the packages will
be published to npm automatically. If you're not ready to do a release
yet, that's fine, whenever you add more changesets to main, this PR will
be updated.


# Releases
## @zus-health/ctw-component-library@1.45.5

### Patch Changes

-   cf13358b: Update the loinc codes for more trending features.
-   1bc578eb: Add support for more observation value types.

Co-authored-by: github-actions[bot] <github-actions[bot]@users.noreply.github.com>