Cdev 287/open diagnostic reports from drawer (#1049)
[CDEV-287](https://zeushealth.atlassian.net/browse/CDEV-287)

[Slack discussion on DR / observations / trend data
referencing](https://zushealth.slack.com/archives/C05231F29LP/p1690564508859959)

This PR has the ability to link observations and open the respective
drawer to that observation. The observation is bolded and easily
identifiable. The drawer got a facelift with a bunch of styling changes.
There is a slight vibration for when you click on the observation which
will be a follow-up ticket.

Additionally, this PR has refactored solutions for the two tickets below
(which was necessary because of the refactoring in this ticket):

1. [CDEV-310: Remove incorrectly coded glucose results from
trending](https://zeushealth.atlassian.net/browse/CDEV-310)
2. [CDEV-350: Contained Resources not working w/ new Diagnostics Report
Drawer](https://zeushealth.atlassian.net/browse/CDEV-350)

![Screenshot 2023-08-02 at 9 25 34
AM](https://github.com/zeus-health/ctw-component-library/assets/98342697/2f747d08-a1bd-4609-98db-57b5aa715269)


[CDEV-287]:
https://zeushealth.atlassian.net/browse/CDEV-287?atlOrigin=eyJpIjoiNWRkNTljNzYxNjVmNDY3MDlhMDU5Y2ZhYzA5YTRkZjUiLCJwIjoiZ2l0aHViLWNvbS1KU1cifQ

---------

Co-authored-by: Priyank Madria <98342697+pmadria3@users.noreply.github.com>
Co-authored-by: Priyank Madria <pmadria@zushealth.com>
Co-authored-by: Patrick Valle <github@patrickvalle.com>