Version Packages (#1068)
This PR was opened by the [Changesets
release](https://github.com/changesets/action) GitHub action. When
you're ready to do a release, you can merge this and the packages will
be published to npm automatically. If you're not ready to do a release
yet, that's fine, whenever you add more changesets to main, this PR will
be updated.


# Releases
## @zus-health/ctw-component-library@1.47.2

### Patch Changes

- c69c3cf5: Improve appearance of ZAP's "More" button, disallowing
wrapping and enforcing font size consistency with the tabs.

Co-authored-by: github-actions[bot] <github-actions[bot]@users.noreply.github.com>