Version Packages (#1100)
This PR was opened by the [Changesets
release](https://github.com/changesets/action) GitHub action. When
you're ready to do a release, you can merge this and the packages will
be published to npm automatically. If you're not ready to do a release
yet, that's fine, whenever you add more changesets to main, this PR will
be updated.


# Releases
## @zus-health/ctw-component-library@1.55.0

### Minor Changes

- 69338e37: Update date range and remove type field for encounters
details drawer
- b66d2fc7: Filter out 'empty' clinical notes from Encounters & Notes
component
- 63e3cfa0: Add functionality to Encounters & Notes table, including
unread notifactions and the ability to dismiss or mark as viewed.
Additionally, makes the "Mark as Viewed" button appear disabled to Zus
employees.

### Patch Changes

-   97528b07: Add speciality to the encounter provider column.

Co-authored-by: github-actions[bot] <github-actions[bot]@users.noreply.github.com>