Version Packages (#1120)
This PR was opened by the [Changesets
release](https://github.com/changesets/action) GitHub action. When
you're ready to do a release, you can merge this and the packages will
be published to npm automatically. If you're not ready to do a release
yet, that's fine, whenever you add more changesets to main, this PR will
be updated.


# Releases
## @zus-health/ctw-component-library@1.56.2

### Patch Changes

- 55e9da0b: Resolved issue with encounters showing "no data" before data
was requested
- 14f93046: Updated "Add Filters" button to take up as much space as the
dropdown menu it opens

Co-authored-by: github-actions[bot] <github-actions[bot]@users.noreply.github.com>
Co-authored-by: Patrick Valle <patrickvalle@users.noreply.github.com>