Version Packages (#1129)
This PR was opened by the [Changesets
release](https://github.com/changesets/action) GitHub action. When
you're ready to do a release, you can merge this and the packages will
be published to npm automatically. If you're not ready to do a release
yet, that's fine, whenever you add more changesets to main, this PR will
be updated.


# Releases
## @zus-health/ctw-component-library@1.57.2

### Patch Changes

- 444b9953: Allow the unread encounters notification icon to show up
outside of the components.
- aa1e7b1f: Address issue where observations with a value of 0 do not
render

Co-authored-by: github-actions[bot] <github-actions[bot]@users.noreply.github.com>