CTW-1348/Embedded ZAP custom fonts (#1138)
The theme passed to the CTWProvider now allows for an optional `iframe`
attribute which includes theming that can't be accomplished via tailwind
as it is currently. The `iframe` theme has 3 optional properties
fontFamily, fontSize and lineHeight which can be used to make the iframe
embedded ZAP look more like the rest of the webapp.