Version Packages (#1150)
This PR was opened by the [Changesets
release](https://github.com/changesets/action) GitHub action. When
you're ready to do a release, you can merge this and the packages will
be published to npm automatically. If you're not ready to do a release
yet, that's fine, whenever you add more changesets to main, this PR will
be updated.


# Releases
## @zus-health/ctw-component-library@1.60.4

### Patch Changes

- e1dd6592: Display notes more securely and scroll upon overflow in a
note.
-   51d268af: Expose use-ctw hook.
-   069f5c3b: Increase security for displaying CCDA docs.

Co-authored-by: github-actions[bot] <github-actions[bot]@users.noreply.github.com>