CTW-1370/iframe ZAP onRecordSave for medications (#1163)
- Handle callbacks for `onAddToRecord` handlers when medications are
added within the embedded iframe ZAP.
- Typing for the ZAP MessageEvents

---
Because we're continuing to add more event types with different use
cases and bodies, this PR also adds some typing to the messages being
sent from parent and iframe. The TLDR is that in there is now a
`payload` and `error` attribute on the message data:
```ts
data: {
  type: string.   // this is unchanged from how we did it before
  payload: T      // this is new, instead of adding attributes to the flat `data` object. A successful
                  //  event can post a message with attributes on the `data.payload` object.
  error?: string  // An unsuccessful event can post an error message.
```
This isn't just stylistic, the reason here is to separate transport
information from the events message and data. The unique structure of a
message should all live in the `payload`. A message sent like
```ts
postMessage({
  type: "ZusAggregatedProfileOnResourceSave",
  error: error.message,
})
```
would convey an error from the `IFrameZusAggregatedProfile` component
itself saying it was unable to handle a
`"ZusAggregatedProfileOnResourceSave"` message. However, a message sent
like
```ts
postMessage({
  type: "ZusAggregatedProfileOnResourceSave",
  payload: {
    resource: fhirResource,
    action: "create",
    error: "Some error message",
  }
})
```
would convey that we have a message about an error that happened in a
sub-component. The message is valid, the payload is valid, it just so
happens that it is informing the other end to pass that "error" into a
callback. The takeaway is that `.error` and `.payload.error` mean
completely different things.

This change won't break anything because in ctw-component-library, only
the new `onAddToRecord` needs to read payload and on the ctw side there
would only be a breaking change if someone uses this version of the
library before we deploy [this ctw
PR](https://github.com/zeus-health/ctw/pull/760) which will make the
standalone ZAP be able to handle both styles of messages, and we can let
that live for a good amount of time.