Version Packages (#1199)
This PR was opened by the [Changesets
release](https://github.com/changesets/action) GitHub action. When
you're ready to do a release, you can merge this and the packages will
be published to npm automatically. If you're not ready to do a release
yet, that's fine, whenever you add more changesets to main, this PR will
be updated.


# Releases
## @zus-health/ctw-component-library@1.65.10

### Patch Changes

- f7e85a43: Fix ADT table empty message and animation for opening
drawers.
-   d19aab56: Transpose wide tables in drawer for better readability
- 889464f1: Only show one date if encounter start and end dates are the
same

Co-authored-by: github-actions[bot] <github-actions[bot]@users.noreply.github.com>