CTW-1393 Search UI improvements (#1203)
- Modified search to only use keyword search, limit to 10 results, and
don't search observations. We will continue to tweak this but these
settings have yielded sensible results so far during testing.
- Created generic SearchResult component and removed custom components
for conditions/documents/etc.
- Modified what data we show for the different search results based on
feedback during meeting with design. Will continue to iterate on these.
- Added empty states.
- Updated styling and made it so the heading is the button to open the
drawer. Still struggling a bit with getting the text overflow working on
the medications heading, but may punt on that in the interest of getting
this out faster for people to validate search results.

---------

Co-authored-by: Michael Rosata <mrosata@zushealth.com>