Version Packages (#1210)
This PR was opened by the [Changesets
release](https://github.com/changesets/action) GitHub action. When
you're ready to do a release, you can merge this and the packages will
be published to npm automatically. If you're not ready to do a release
yet, that's fine, whenever you add more changesets to main, this PR will
be updated.


# Releases
## @zus-health/ctw-component-library@1.65.13

### Patch Changes

-   856d6bae: Show discharge disposition in encounter table if available
-   d37007a6: Adds default sort and sorting settings to the ADT table

Co-authored-by: github-actions[bot] <github-actions[bot]@users.noreply.github.com>