CTW-1398 Remove version history from resource history drawer (#1222)
We have deemed that it is unnecessary to surface the version
history *of a single FHIR resource* in the history section of the
details drawer. The primary intent of this section is to highlight what
other organizations have participated in care, so being able to show
something like the status transitions of a condition (eg. when it went
from active to inactive) is not necessary. In an effort to simplify the
code and the logic behind this feature, reduce calls to the back-end,
and address the bug in CTW-1398 where we're showing "Unknown" as the
title for older versions of a condition resource, we are removing this
functionality.