CTW-1388 Enable semantic and generative search results (#1231)
Updating search so that we include the semantic search if the search
phrase has more than 2 words. If the user ends with a question mark
we'll also include the AI generated response. This behavior is subject
to change in the future, just doing this as a short term way of making
it available in the app.

Also fixing an issue with how we retrieve documents from FQS by ID.