CTW-1403 Moving search to a tab in the ZAP (#1232)
Moving search out of the ZAP Title bar (which is typically disabled and
we may even want to consider removing altogether) and into a dedicated
tab. This kind of hides it in the "More" menu but that's OK for the time
being while it's feature flagged and experimental. UX is going to work
on a longer term plan for this feature while we use this as a way of
simply getting it available in the ZAP.