Version Packages (#1236)
This PR was opened by the [Changesets
release](https://github.com/changesets/action) GitHub action. When
you're ready to do a release, you can merge this and the packages will
be published to npm automatically. If you're not ready to do a release
yet, that's fine, whenever you add more changesets to main, this PR will
be updated.


# Releases
## @zus-health/ctw-component-library@1.67.2

### Patch Changes

- a17eb193: Fix use patient query by only running it if patientId is
provided.
-   f4dab1c5: Code refactoring to make resource drawers more extensible.
-   71d1bb55: Add diagnostic reports to search results.
-   34bd51dc: Move search to a tab in the ZAP.
-   934f9696: Enable semantic and generative AI search results.

Co-authored-by: github-actions[bot] <github-actions[bot]@users.noreply.github.com>