CTW-1404 Refactoring diagnostic report queries (#1252)
We are currently making some very expensive calls to retrieve diagnostic
reports. Every time we load the ZAP (in order to show the red dot
indicating there are unread diagnostic reports) we retrieve up to 1000
observations and up to 1000 diagnostic reports with references to
patients/practitioner/observations/medication-requests, and we do that
second query TWICE!

This PR refactors our diagnostic reports data retrieval into 2 hooks -
`usePatientDiagnosticReports` with options for a limit and whether or
not you include observations. We will use this hook with a limit of 20
and not including observations when retrieving data to inform the red
dot. Basically this means that if any of the 20 most recently updated
resources are unread then show the red dot. This feels like a
good-enough answer for "is there any data for me to look at".

The second hook is `usePatientDiagnosticReportsWithTrendData` which
takes in a limit as well. This will query both observations and
diagnostic reports and is still somewhat expensive, but at least now
we're only querying diagnostic report data once. I left in a lot of the
logic that re-constructs DiagnosticReportModels with observation trend
arrays which themselves contain DiagnosticReportModels as links to their
parent diagnostic reports. That's all pretty intensive and could maybe
use refactoring in the future, but for now at least the network calls
are less intensive.

This PR also refactors out some FQS feature flagging that's no longer
necessary and reduces our FQS query limit to 500 from 1000.